Rename default_dtype.py to default_type.py
