+++ e3nn/util/__init__.py
@@ -1,3 +1,3 @@
 from .eval_code import *  # noqa
-from .default_dtype import *  # noqa
+from .default_type import *  # noqa
 from .broadcast import *  # noqa
